Koblet med backend
- added in backend: files for connecting to database.
connect.js to connect to the database. /controller/tjenesteeier.js for sql queries, /routes/tjenesteeier for connect queries with database. index.js as connection point.
env.example for showing how to connect to database
added files in react: Tjenesteeier.jsx for showing and deleting a tjenesteeier. add.jsx for adding a new tjenesteeier
- changed: app.jsx for routing tjenesteeier.